requirements.txt
flask
anthropic
python-dotenv
requests
pypdf
python-docx
rank-bm25
duckduckgo-search
